Field map correction
This script accepts as an input the standard functional and anatomical images. From the EPI series it accepts a magnitude image (which does not need to be skull stripped) and the phase images. The skull stripping is done with a high threshold to ensure that it is tight around the edges. The extracted magnitude image is now sent for the field map to be prepared (fsl_prepare_fieldmap). The prepared fieldmap will now be used for performing epi distortion correction and co-registration.